// src/common/enums/amenity.enum.ts
export enum AmenityCondition {
  BROKEN = "broken",
  POOR = "poor",
  GOOD = "good",
  EXCELLENT = "excellent",
}

export enum AmenityCategory {
  FINISHING = "finishing", // e.g., "Marble Countertops", "Porcelain Tiles"
  FIXED_APPLIANCE = "fixed_appliance", // e.g., "Cooktop", "Oven"
  CUSTOM_FURNITURE = "custom_furniture", // e.g., "Kitchen Cabinets", "Closet"
  INFRASTRUCTURE = "infrastructure", // e.g., "Air Conditioning", "Solar Heating"
  LEISURE = "leisure", // e.g., "Pool", "BBQ Grill"
  MOVABLE_FURNITURE = "movable_furniture", // e.g., "Sofa", "Bed"
}